Changes to fix cell- versus point-centred plotting
Image pixels are cell-centered values. Data plots
are point-centered values. To ensure that there is
no shifting when bilinear interpolation toggled on/off,
and to ensure that image pixels are correctly mapped
to data pixels and world coords, we must be careful
to account for centre- versus point-centred data
conventions. This patch is an attempt to do that.